Issue #106: Incorrect data returned by cltr::get_enabled_plugins() causes uses unit test failure
The function get_enabled_plugins() is used by the plugin manager to return the plugin names as an array containing $pluginname => $pluginname.For this reason the getting plugin instances is moved in a separate function.
Add unit tests to test cltr::get_enabled_plugins() and cltr::get_enabled_plugin_instances().